+++ ignite/services/scaffolder/module.go
@@ -165,12 +165,17 @@ func (s Scaffolder) CreateModule(
 	moduleName = mfName.LowerCase
 
 	// Check if the module name is valid
+<<<<<<< HEAD
 	if err := checkModuleName(s.path, moduleName); err != nil {
 		return sm, err
+=======
+	if err := checkModuleName(s.appPath, moduleName); err != nil {
+		return err
+>>>>>>> 6364ecbf (feat: support custom proto path (#4071))
 	}
 
 	// Check if the module already exist
-	ok, err := moduleExists(s.path, moduleName)
+	ok, err := moduleExists(s.appPath, moduleName)
 	if err != nil {
 		return sm, err
 	}
@@ -191,8 +196,13 @@ func (s Scaffolder) CreateModule(
 	}
 
 	// Check dependencies
+<<<<<<< HEAD
 	if err := checkDependencies(creationOpts.dependencies, s.path); err != nil {
 		return sm, err
+=======
+	if err := checkDependencies(creationOpts.dependencies, s.appPath); err != nil {
+		return err
+>>>>>>> 6364ecbf (feat: support custom proto path (#4071))
 	}
 
 	opts := &modulecreate.CreateOptions{
@@ -200,7 +210,8 @@ func (s Scaffolder) CreateModule(
 		ModulePath:   s.modpath.RawPath,
 		Params:       params,
 		AppName:      s.modpath.Package,
-		AppPath:      s.path,
+		AppPath:      s.appPath,
+		ProtoDir:     s.protoDir,
 		IsIBC:        creationOpts.ibc,
 		IBCOrdering:  creationOpts.ibcChannelOrdering,
 		Dependencies: creationOpts.dependencies,