+++ ignite/templates/app/options.go
@@ -4,6 +4,7 @@ package app
 type Options struct {
 	AppName          string
 	AppPath          string
+	ProtoDir         string
 	GitHubPath       string
 	BinaryNamePrefix string
 	ModulePath       string