+++ ignite/templates/typed/singleton/singleton.go
@@ -105,7 +105,7 @@ const (
 //   - Existence of a service with name "Query". Adds the rpc's there.
 func protoRPCModify(opts *typed.Options) genny.RunFn {
 	return func(r *genny.Runner) error {
-		path := opts.ProtoPath("query.proto")
+		path := opts.ProtoFile("query.proto")
 		f, err := r.Disk.Find(path)
 		if err != nil {
 			return err
@@ -192,7 +192,7 @@ func clientCliQueryModify(replacer placeholder.Replacer, opts *typed.Options) ge
 //   - Existence of a message with name "GenesisState". Adds the field there.
 func genesisProtoModify(opts *typed.Options) genny.RunFn {
 	return func(r *genny.Runner) error {
-		path := opts.ProtoPath("genesis.proto")
+		path := opts.ProtoFile("genesis.proto")
 		f, err := r.Disk.Find(path)
 		if err != nil {
 			return err
@@ -364,7 +364,7 @@ if found {
 //   - A service named "Msg" to exist in the proto file, it appends the RPCs inside it.
 func protoTxModify(opts *typed.Options) genny.RunFn {
 	return func(r *genny.Runner) error {
-		path := opts.ProtoPath("tx.proto")
+		path := opts.ProtoFile("tx.proto")
 		f, err := r.Disk.Find(path)
 		if err != nil {
 			return err